+++ package.json
@@ -34,7 +34,7 @@
     "lint": "eslint .",
     "test": "npm run test:unit && npm run test:typescript",
     "test:unit": "node scripts/unit.js",
-    "test:typescript": "tsd",
+    "test:typescript": "tstyche",
     "performance": "npm run lint && node performanceTest/test",
     "preversion": "npm run test && git push",
     "postversion": "git push && git push --tags"
@@ -54,7 +54,7 @@
     "fastify": "^5.6.2",
     "neostandard": "^0.12.2",
     "serverless-http": "^4.0.0",
-    "tsd": "^0.33.0"
+    "tstyche": "^7.0.0"
   },
   "publishConfig": {
     "access": "public"
